server-python/requirements.txt
blinker==1.7.0
chardet==5.2.0
click==8.1.7
flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
marshmallow==3.20.2
mysql-connector-python==8.0.28
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
PyMySQL==1.1.0
pytz==2023.3.post1
reportlab==4.0.9
SQLAlchemy==2.0.25
typing-extensions==4.9.0
werkzeug==3.0.1
zipp==3.17.0
